zonepb: dedupe some validations for zone configs between schema changers
Epic: none

Release note: None
